+++ src/Paket.VisualStudio/Restore/PaketRestorer.cs
@@ -1,0 +1,17 @@
+using System.Collections.Generic;
+using System.Linq;
+using Microsoft.FSharp.Collections;
+using Microsoft.FSharp.Core;
+
+namespace Paket.VisualStudio.Restore
+{
+    public class PaketRestorer : IPackageRestorer
+    {
+        public void Restore(Dependencies dependencies, IEnumerable<RestoringProject> project)
+        {
+            dependencies.Restore(
+                FSharpOption<string>.None,
+                ListModule.OfSeq(project.Select(p => p.ReferenceFile)));
+        }
+    }
+}