+++ src/Paket.VisualStudio/SolutionExplorer/PaketMenuCommandService.cs
@@ -22,6 +22,7 @@ namespace Paket.VisualStudio.SolutionExplorer
     {
         public string DependenciesFileName { get; set; }
         public string ReferencesFileName { get; set; }
+        public string GroupName { get; set; }
         public string PackageName { get; set; }
     }
 
@@ -50,6 +51,7 @@ namespace Paket.VisualStudio.SolutionExplorer
         private void RegisterCommands()
         {
             RegisterCommand(CommandIDs.UpdatePackage, UpdatePackage, null);
+            RegisterCommand(CommandIDs.UpdateGroup, UpdateGroup, null);
             RegisterCommand(CommandIDs.RemovePackage, RemovePackage, OnlyBelowDependenciesFileNodes);
             RegisterCommand(CommandIDs.RemovePackageFromProject, RemovePackageFromProject, OnlyBelowReferencesFileNodes);
             RegisterCommand(CommandIDs.CheckForUpdates, CheckForUpdates, OnlyDependenciesFileNodes);
@@ -186,10 +188,11 @@ namespace Paket.VisualStudio.SolutionExplorer
             var info = new PackageInfo();
             info.DependenciesFileName = node.Id.GetFileName();
             info.PackageName = node.GetPackageName();
+            info.GroupName = node.GetGroupName();
 
             var projectGuids =
                     Paket.Dependencies.Locate(info.DependenciesFileName)
-                        .FindProjectsFor(info.PackageName)
+                        .FindProjectsFor(info.GroupName,info.PackageName)
                         .Select(project => project.GetProjectGuid())
                         .ToArray();
 
@@ -271,6 +274,7 @@ namespace Paket.VisualStudio.SolutionExplorer
             info.DependenciesFileName = node.Id.GetFileName();
             info.ReferencesFileName = node.Id.GetFileName();
             info.PackageName = node.GetPackageName();
+            info.GroupName = node.GetGroupName();
 
             System.Threading.Tasks.Task.Run(() =>
             {
@@ -322,7 +326,7 @@ namespace Paket.VisualStudio.SolutionExplorer
 
         private void Restore(object sender, EventArgs e)
         {
-            RunCommand("paket-restore.html", info => // Do we need to unload?
+            RunCommandAndReloadAllProjects("paket-restore.html", info =>
             {
                 Paket.Dependencies.Locate(tracker.GetSelectedFileName())
                     .Restore();
@@ -343,7 +347,17 @@ namespace Paket.VisualStudio.SolutionExplorer
             RunCommandOnPackageAndReloadAllDependendProjects("paket-update.html#Updating-a-single-package", info =>
             {
                 Paket.Dependencies.Locate(info.DependenciesFileName)
-                    .UpdatePackage(info.PackageName, Microsoft.FSharp.Core.FSharpOption<string>.None, false, false);
+                    .UpdatePackage(Microsoft.FSharp.Core.FSharpOption<string>.Some(info.GroupName), info.PackageName, Microsoft.FSharp.Core.FSharpOption<string>.None, false, false, SemVerUpdateMode.NoRestriction);
+            });
+        }
+        
+
+        private void UpdateGroup(object sender, EventArgs e)
+        {
+            RunCommandOnPackageAndReloadAllDependendProjects("paket-update.html#Updating-a-single-group", info =>
+            {
+                Paket.Dependencies.Locate(info.DependenciesFileName)
+                    .UpdateGroup(info.GroupName, false, false, false, false, true, SemVerUpdateMode.NoRestriction);
             });
         }
 
@@ -352,7 +366,7 @@ namespace Paket.VisualStudio.SolutionExplorer
             RunCommandOnPackageAndReloadAllDependendProjects("paket-remove.html", info =>
             {
                 Paket.Dependencies.Locate(info.DependenciesFileName)
-                    .Remove(info.PackageName);
+                    .Remove(Microsoft.FSharp.Core.FSharpOption<string>.Some(info.GroupName), info.PackageName);
             });
         }
 
@@ -397,7 +411,7 @@ namespace Paket.VisualStudio.SolutionExplorer
             RunCommandOnPackageInUnloadedProject("paket-remove.html#Removing-from-a-single-project", info =>
             {
                 Paket.Dependencies.Locate(info.DependenciesFileName)
-                    .RemoveFromProject(info.PackageName, false, false, info.ReferencesFileName, true);
+                    .RemoveFromProject(Microsoft.FSharp.Core.FSharpOption<string>.Some(info.GroupName), info.PackageName, false, false, info.ReferencesFileName, true);
             });
         }